chore: Update User and Post models
This commit updates the User and Post models by adding a new field  to the User model and  field to the Post model. The  field is of type TEXT and is required, while the  field is also of type TEXT and is nullable. These changes were made to accommodate the requirements of the application.

Complete basic api blogs query